--- src/main/java/nextstep/ladder/Players.java
@@ -1,11 +1,0 @@
-package nextstep.ladder;
-
-import java.util.List;
-
-public class Players {
-    private List<Player> players;
-
-    public Players(List<Player> players) {
-        this.players = players;
-    }
-}
